requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
mcstatus>=11.0.0
python-a2s
